developpement/MNIST/requirements_gradio.txt
pillow==10.3.0
gradio==5.49
numpy==1.26.4
